Change BUILD_ANTMICRO to BUILD_UPSTREAM, split functions with define to different file
